pyproject.toml
[tool.poetry]
name = "diary_project"
version = "0.1.0"
description = "Online diary app"
authors = ["NovaHFly <mugiviska@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
djoser = "^2.3.1"
django-filter = "^25.1"


[tool.poetry.group.testing.dependencies]
ruff = "^0.11.4"
pytest = "^8.3.5"
pytest-django = "^4.11.1"
pytest-lazy-fixtures = "^1.1.2"

[tool.pytest.ini_options]
pythonpath = "diary_project/"
DJANGO_SETTINGS_MODULE = "diary_project.settings"
testpaths = "tests/"
python_files = "test_*.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
